refactor: put Analysis Cohort logic behind a shared boundary
Analysis Cohort semantics now live in shared/core/cohorts/analysis so server routes and client shims use the same domain implementation. Client Cohort Definition IDs now match the backend UUID contract.

Constraint: Preserve the legacy CohortEngine public symbol as the Exit Cohort Model surface.

Rejected: Merge Analysis Cohorts into CohortEngine | it would conflate performance grouping with exit/value progression modeling.

Rejected: Repo-wide server-to-client import ban | useful but broader than this regression.

Confidence: high

Scope-risk: moderate

Directive: Server code must import Analysis Cohort logic from shared/core/cohorts/analysis, not client/src/core/cohorts.

Tested: npx vitest run tests/unit/cohorts/analyze-cohorts.test.ts tests/unit/server/cohort-routes-registration.test.ts tests/unit/contract/cohort-analysis-boundary.test.ts tests/unit/contract/funds-boundary-guard.test.ts --project=server

Tested: npm run check

Not-tested: Full suite deferred until final verification.
