+++ server/routes/cohort-analysis.ts
@@ -25,6 +25,7 @@ import {
   investmentLots,
 } from '@shared/schema';
 import { CohortAnalyzeRequestSchema, CohortAnalyzeResponseSchema } from '@shared/types';
+import { analyzeCohorts } from '@shared/core/cohorts/analysis/advanced-engine';
 import { normalizeSector, slugifySector } from '@shared/utils/sector-normalization';
 
 const router = express.Router();
@@ -187,9 +188,6 @@ router['post']('/analyze', async (req: Request, res: Response) => {
       .from(investmentLots)
       .innerJoin(investments, eq(investmentLots.investmentId, investments.id))
       .where(eq(investments.fundId, fundId));
-
-    // Import and run the analysis engine dynamically to avoid circular imports
-    const { analyzeCohorts } = await import('../../client/src/core/cohorts/advanced-engine.js');
 
     const response = analyzeCohorts({
       request,
